+++ python/cog/predictor.py
@@ -31,7 +31,7 @@ except ImportError:  # Python < 3.8
     from typing_compat import get_args, get_origin  # type: ignore
 
 import yaml
-from pydantic import ConfigDict, BaseModel, Field, create_model
+from pydantic import BaseModel, ConfigDict, Field, create_model
 from pydantic.fields import FieldInfo
 
 # Added in Python 3.9. Can be from typing if we drop support for <3.9
