+++ python/cog/types_shared.py
@@ -1,12 +1,9 @@
 import io
 import mimetypes
 import os
-import pathlib
-import shutil
-import tempfile
 import urllib.parse
 import urllib.request
-from typing import Any, Dict, Iterator, Optional
+from typing import Any, Dict, Iterator
 
 import requests
 
@@ -17,45 +14,6 @@ FILENAME_ILLEGAL_CHARS = set("\u0000/")
 # tempfile.NamedTemporaryFile, etc.
 FILENAME_MAX_LENGTH = 200
 
-
-class URLPath(pathlib.PosixPath):
-    """
-    URLPath is a nasty hack to ensure that we can defer the downloading of a
-    URL passed as a path until later in prediction dispatch.
-
-    It subclasses pathlib.PosixPath only so that it can pass isinstance(_,
-    pathlib.Path) checks.
-    """
-
-    _path: Optional[Path]
-
-    def __init__(self, *, source: str, filename: str, fileobj: io.IOBase) -> None:
-        self.source = source
-        self.filename = filename
-        self.fileobj = fileobj
-
-        self._path = None
-
-    def convert(self) -> Path:
-        if self._path is None:
-            dest = tempfile.NamedTemporaryFile(suffix=self.filename, delete=False)
-            shutil.copyfileobj(self.fileobj, dest)
-            self._path = Path(dest.name)
-        return self._path
-
-    def unlink(self, missing_ok: bool = False) -> None:
-        if self._path:
-            # TODO: use unlink(missing_ok=...) when we drop Python 3.7 support.
-            try:
-                self._path.unlink()
-            except FileNotFoundError:
-                if not missing_ok:
-                    raise
-
-    def __str__(self) -> str:
-        # FastAPI's jsonable_encoder will encode subclasses of pathlib.Path by
-        # calling str() on them
-        return self.source
 
 
 class URLFile(io.IOBase):
