+++ python/cog/types_v1.py
@@ -1,6 +1,4 @@
 import io
-import mimetypes
-import os
 import pathlib
 import shutil
 import tempfile
@@ -8,9 +6,10 @@ import urllib.parse
 import urllib.request
 from typing import Any, Dict, Iterator, List, Optional, TypeVar, Union
 
-import requests
 from pydantic import Field, SecretStr
-from .types_shared import URLFile, URLPath, get_filename
+
+from .types_shared import URLFile, get_filename
+
 
 def Input(
     default: Any = ...,
@@ -103,6 +102,46 @@ class Path(pathlib.PosixPath):
         """Defines what this type should be in openapi.json"""
         # https://json-schema.org/understanding-json-schema/reference/string.html#uri-template
         field_schema.update(type="string", format="uri")
+
+
+class URLPath(pathlib.PosixPath):
+    """
+    URLPath is a nasty hack to ensure that we can defer the downloading of a
+    URL passed as a path until later in prediction dispatch.
+
+    It subclasses pathlib.PosixPath only so that it can pass isinstance(_,
+    pathlib.Path) checks.
+    """
+
+    _path: Optional[Path]
+
+    def __init__(self, *, source: str, filename: str, fileobj: io.IOBase) -> None:
+        self.source = source
+        self.filename = filename
+        self.fileobj = fileobj
+
+        self._path = None
+
+    def convert(self) -> Path:
+        if self._path is None:
+            dest = tempfile.NamedTemporaryFile(suffix=self.filename, delete=False)
+            shutil.copyfileobj(self.fileobj, dest)
+            self._path = Path(dest.name)
+        return self._path
+
+    def unlink(self, missing_ok: bool = False) -> None:
+        if self._path:
+            # TODO: use unlink(missing_ok=...) when we drop Python 3.7 support.
+            try:
+                self._path.unlink()
+            except FileNotFoundError:
+                if not missing_ok:
+                    raise
+
+    def __str__(self) -> str:
+        # FastAPI's jsonable_encoder will encode subclasses of pathlib.Path by
+        # calling str() on them
+        return self.source
 
 
 Item = TypeVar("Item")
